refactor: separate styles, related component
