+++ site/src/components/input/Select/check_icon.tsx
@@ -1,0 +1,5 @@
+export const Check = (props: Record<string, unknown>) => (
+  <svg {...props} xmlns="http://www.w3.org/2000/svg" width="24" height="24">
+    <path d="m10 15.586-3.293-3.293-1.414 1.414L10 18.414l9.707-9.707-1.414-1.414z" />
+  </svg>
+)
